+++ auth-service/src/main/java/com/logistics/platform/auth_service/application/service/UserService.java
@@ -1,7 +1,7 @@
 package com.logistics.platform.auth_service.application.service;
 
-import com.logistics.platform.auth_service.application.dto.UserResDto;
-import com.logistics.platform.auth_service.common.exception.CustomApiException;
+import com.logistics.platform.auth_service.presentation.response.UserResDto;
+import com.logistics.platform.auth_service.presentation.global.exception.CustomApiException;
 import com.logistics.platform.auth_service.domain.model.User;
 import com.logistics.platform.auth_service.domain.repository.UserRepository;
 import com.logistics.platform.auth_service.presentation.request.DeleteReqDto;
