+++ auth-service/src/main/java/com/logistics/platform/auth_service/presentation/global/exception/CustomApiException.java
@@ -1,4 +1,4 @@
-package com.logistics.platform.auth_service.common.exception;
+package com.logistics.platform.auth_service.presentation.global.exception;
 
 public class CustomApiException extends RuntimeException{
     public CustomApiException(String message) {
